+++ mipac/config.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Literal, TypedDict
+from typing import Literal
 
 
 @dataclass
@@ -26,6 +26,7 @@ class Config:
         is_ayuskey: bool = False,
         use_version: Literal[13, 12, 11] = 12,
         cache: CacheConfigData | None = None,
+        use_version_autodetect: bool = True,
     ) -> None:
         self.distro: IMisskeyDistribution = distro
         self.is_ssl: bool = is_ssl
@@ -33,6 +34,7 @@ class Config:
         self.is_ayuskey: bool = is_ayuskey
         self.use_version: Literal[13, 12, 11] = use_version
         self.cache: CacheConfig = CacheConfig(cache or CacheConfigData())
+        self.use_version_autodetect: bool = use_version_autodetect
 
     def from_dict(
         self,
@@ -41,6 +43,7 @@ class Config:
         is_ayuskey: bool | None = None,
         use_version: Literal[13, 12, 11] | None = None,
         cache: CacheConfigData | None = None,
+        use_version_autodetect: bool | None = None,
     ):
         self.host = host or self.host
         self.is_ssl = is_ssl if is_ssl is not None else self.is_ssl
@@ -48,6 +51,7 @@ class Config:
         self.use_version = use_version or self.use_version
         if cache:
             self.cache = CacheConfig(cache)
+        self.use_version_autodetect = use_version_autodetect or self.use_version_autodetect
 
 
 config = Config()